Option 2:
Just inherit from dict.
Advantage: No object breakage due to name collisions, when storing in the dict.
Disadvantage: Dict introduces new name pollution with names which you may well want to use in PySON, such as the name "update".
(not a polished solution, just for history)

Keeping this in a commit, but planning to move on to Option 3: Inherit from dict, but rename the members via a metaclass, adding a configurable prefix, so that collisions can be easily avoided per application.